split array metafunctions into bits
