+++ include/nmtools/meta/array.hpp
@@ -5,701 +5,26 @@
 #include "nmtools/meta/traits.hpp"
 #include "nmtools/meta/loop.hpp"
 
+#include "nmtools/meta/bits/transform/remove_cvref.hpp"
+
+#include "nmtools/meta/bits/array/fixed_dim.hpp"
+#include "nmtools/meta/bits/array/fixed_index_array_size.hpp"
+#include "nmtools/meta/bits/array/fixed_ndarray_shape.hpp"
+#include "nmtools/meta/bits/array/hybrid_index_array_max_size.hpp"
+#include "nmtools/meta/bits/array/hybrid_ndarray_max_size.hpp"
+#include "nmtools/meta/bits/array/nested_array_dim.hpp"
+#include "nmtools/meta/bits/array/nested_array_size.hpp"
+#include "nmtools/meta/bits/array/remove_all_nested_array_dim.hpp"
+#include "nmtools/meta/bits/array/remove_nested_array_dim.hpp"
+#include "nmtools/meta/bits/array/resize_hybrid_index_array_max_size.hpp"
+#include "nmtools/meta/bits/array/resize_hybrid_ndarray_dim.hpp"
+#include "nmtools/meta/bits/array/resize_hybrid_ndarray_max_size.hpp"
+
 #include <type_traits>
 #include <tuple>
 
-namespace nmtools::meta
-{
-    /**
-     * @addtogroup meta
-     * @{
-     */
-
-    /**
-     * @brief returns the number of elements of fixed shape at its 1st axis.
-     *
-     * @tparam T type to check
-     * @tparam typename customization point
-     */
-    template <typename T, typename=void>
-    struct nested_array_size
-    {
-        static constexpr auto value = 0;
-    }; // nested_array_size
-
-    /**
-     * @brief specialization of nested_array_size for type that has tuple_size and value_type.
-     * 
-     * @tparam T type tot check
-     */
-    template <typename T>
-    struct nested_array_size<T
-        , enable_if_t<has_tuple_size_v<T> && has_value_type_v<T>>
-    > {
-        static constexpr auto value = std::tuple_size_v<T>;
-    }; // nested_array_size
-
-    /**
-     * @brief helper variable template to get the number of elements of fixed shape at its 1st axis.
-     * 
-     * @tparam T type to check
-     * @see nested_array_size
-     */
-    template <typename T>
-    inline constexpr auto nested_array_size_v = nested_array_size<T>::value;
-
-    /**
-     * @brief get the number of dimension of (possibly) nested array.
-     *
-     * By default, check using expr::square_bracket.
-     * 
-     * @tparam T type to check
-     * @tparam typename
-     */
-    template <typename T, typename=void>
-    struct nested_array_dim
-    {
-        static constexpr auto value = 0;
-    }; // nested_array_dim
-
-    /**
-     * @brief specialization of nested_array_dim
-     *
-     * Sepcialized when T square bracket expression with size_t is well-formed,
-     * checked using has_square_bracket. Recursively instantiate nested_array_dim
-     * with decreasing dimension.
-     * 
-     * @tparam T type to check
-     * @see expr::square_bracket
-     * @see has_square_bracket
-     */
-    template <typename T>
-    struct nested_array_dim<T,enable_if_t<has_square_bracket_v<T,size_t>>>
-    {
-        using value_type = std::remove_reference_t<expr::square_bracket<T,size_t>>;
-        static constexpr auto value = 1 + nested_array_dim<value_type>::value;
-    }; // nested_array_dim
-
-    /**
-     * @brief helper variable template for nested_array_dim.
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline constexpr auto nested_array_dim_v = nested_array_dim<T>::value;
-
-    namespace detail
-    {
-        /**
-         * @brief remove N number of dimension of a nested array.
-         * 
-         * @tparam T type to transfrom
-         * @tparam N number of dimension to remove
-         */
-        template <typename T, size_t N, typename=void>
-        struct remove_nested_array_dim
-        {
-            /**
-             * @brief helper function declaration to deduce the nested type of given type U
-             *
-             * Actual case when U is subscriptable.
-             * 
-             * @tparam U array to check
-             * @return enable_if_t<has_square_bracket_v<U,size_t>,expr::square_bracket<U,size_t>> 
-             * @note the deduction is performed as static fn instead of specailization to prevent
-             * ambiguous instantiation with terminal case when N==0
-             */
-            template <typename U>
-            static auto try_nested_type()
-                -> enable_if_t<has_square_bracket_v<U,size_t>,expr::square_bracket<U,size_t>>;
-            /**
-             * @brief helper function declaration to deduce the nested type of given type U
-             * 
-             * Return identity if U is not subscriptable.
-             * 
-             * @tparam U type to check
-             * @return enable_if_t<!has_square_bracket_v<U,size_t>,T> 
-             * @note the deduction is performed as static fn instead of specailization to prevent
-             * ambiguous instantiation with terminal case when N==0
-             */
-            template <typename U>
-            static auto try_nested_type()
-                -> enable_if_t<!has_square_bracket_v<U,size_t>,U>;
-            using nested_type = remove_cvref_t<decltype(try_nested_type<T>())>;
-            using type = typename detail::remove_nested_array_dim<nested_type,N-1>::type;
-        }; // remove_nested_array_dim
-
-        // @note providing this specialization while also providing specializaiton for N==0
-        // will cause error: ambiguous template instantiation
-        // template <typename T, size_t N>
-        // struct remove_nested_array_dim<T,N,std::enable_if_t<has_square_bracket_v<T,size_t>>>
-        // {
-        //     using nested_type = remove_cvref_t<expr::square_bracket<T,size_t>>;
-        //     using type = typename detail::remove_nested_array_dim<nested_type,N-1>::type;
-        // };
-
-        /**
-         * @brief terminal case for remove_nested_array_dim
-         * 
-         * @tparam T 
-         */
-        template <typename T>
-        struct remove_nested_array_dim<T,0,void>
-        {
-            using type = remove_cvref_t<T>;
-        }; // remove_nested_array_dim
-    } // namespace detail
-
-    /**
-     * @brief remove N number of dimension of a nested array.
-     * 
-     * @tparam T type to transform
-     * @tparam N number of dimension to be removed
-     * @tparam typename 
-     */
-    template <typename T, size_t N, typename=void>
-    struct remove_nested_array_dim
-    {
-        using type = T;
-    }; // remove_nested_array_dim
-
-    template <typename T, size_t N>
-    struct remove_nested_array_dim<T,N,enable_if_t<has_square_bracket_v<T,size_t>>>
-    {
-        using type = typename detail::remove_nested_array_dim<T,N>::type;
-    }; // remove_nested_array_dim
-
-    /**
-     * @brief helper alias template to remove N number of dimension of a nested array.
-     * 
-     * @tparam T type to transform
-     * @tparam N number of dimension to be removed
-     */
-    template <typename T, size_t N>
-    using remove_nested_array_dim_t = typename remove_nested_array_dim<T,N>::type;
-
-    /**
-     * @brief Get element type of nested array.
-     * 
-     * If T is some type that has expression t[i][j]...,
-     * provide member typedef with type of that element type with maximum indexing,
-     * otherwise provide typedef with T.
-     * 
-     * 
-     * @tparam T type to check
-     * @tparam typename 
-     */
-    template <typename T, typename=void>
-    struct remove_all_nested_array_dim
-    {
-        using type = T;
-    }; // remove_all_nested_array_dim
-
-    /**
-     * @brief Specialization of remove_all_nested_array_dim.
-     *
-     * Enabled when given type T can be 'sliced' with size_t.
-     * 
-     * @tparam T type to check
-     * @see meta::has_square_bracket
-     */
-    template <typename T>
-    struct remove_all_nested_array_dim<T,enable_if_t<has_square_bracket_v<T,size_t>>>
-    {
-        // recursively call itself until it reach default case
-        using value_type = remove_cvref_t<expr::square_bracket<T,size_t>>;
-        using type = typename remove_all_nested_array_dim<value_type>::type;
-    }; // remove_all_nested_array_dim
-
-    /**
-     * @brief helper alias template for remove_all_nested_array_dim.
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    using remove_all_nested_array_dim_t = typename remove_all_nested_array_dim<T>::type;
-
-    /** @} */ // end group meta
-
-} // namespace nmtools::meta
-
-namespace nmtools::meta
-{
-    /**
-     * @brief check if T t{} are:
-     * - t[0][0][0] is valid
-     * 
-     * @tparam T 
-     * @tparam void 
-     */
-    template <typename T, typename=void>
-    struct is_array3d
-    {
-        static constexpr inline auto _check()
-        {
-            if constexpr (has_atnd_v<T,size_t,size_t,size_t>)
-                return true;
-            else if constexpr (has_funcnd_v<T,size_t,size_t,size_t>)
-                return true;
-            else if constexpr (has_bracketnd_v<T,size_t,size_t,size_t>)
-                return true;
-            else if constexpr (std::is_array_v<T> && std::rank_v<T> == 3)
-                return true;
-            else if constexpr (nested_array_dim_v<T> == 3)
-                return true;
-            else return false;
-        } // _check
-        static constexpr inline auto value = _check();
-    }; // is_array3d
-
-    template <typename T>
-    inline constexpr bool is_array3d_v = is_array3d<T>::value;
-
-    // TODO: remove
-    /**
-     * @brief check if T t{} are:
-     * - t[0][0] is valid
-     * 
-     * @tparam T 
-     * @tparam void 
-     */
-    template <typename T, typename=void>
-    struct is_array2d
-    {
-        static constexpr inline auto _check()
-        {
-            if constexpr (has_atnd_v<T,size_t,size_t>)
-                return true;
-            else if constexpr (has_funcnd_v<T,size_t,size_t>)
-                return true;
-            else if constexpr (has_bracketnd_v<T,size_t,size_t>)
-                return true;
-            else if constexpr (std::is_array_v<T> && std::rank_v<T> == 2)
-                return true;
-            else if constexpr (nested_array_dim_v<T> == 2)
-                return true;
-            else return false;
-        } // _check
-        static constexpr inline auto value = _check();
-    }; // is_array2d
-
-    template <typename T>
-    inline constexpr bool is_array2d_v = is_array2d<T>::value;
-
-    // TODO: remove
-    /**
-     * @brief check if T t{} are:
-     * - t[0] is valid
-     * 
-     * @tparam T 
-     * @tparam void 
-     */
-    template <typename T, typename=void>
-    struct is_array1d
-    {
-        static constexpr inline auto _check()
-        {
-            if constexpr (has_atnd_v<T,size_t>)
-                return true;
-            else if constexpr (has_funcnd_v<T,size_t>)
-                return true;
-            else if constexpr (has_square_bracket_v<T,size_t>)
-                return true;
-            else if constexpr (std::is_array_v<T> && std::rank_v<T> == 1)
-                return true;
-            else if constexpr (nested_array_dim_v<T> == 1)
-                return true;
-            else return false;
-        } // _check
-        static constexpr inline auto value = _check();
-    }; // is_array1d
-
-    /**
-     * @brief helper variable template to check if T is 1d array-like
-     * 
-     * @tparam T 
-     */
-    template <typename T>
-    inline constexpr bool is_array1d_v = is_array1d<T>::value;
-} // namespace nmtools::meta
-
-namespace nmtools::meta
-{
-    /**
-     * @addtogroup traits
-     * @{
-     */
-
-    /**
-     * @brief get fixed-array size at compile-time.
-     * well-formed specialization should have `value` and `value_type`.
-     * 
-     * @tparam T type to check
-     * @tparam typename optional for sfinae
-     */
-    template <typename T, typename=void>
-    struct fixed_ndarray_shape
-    {
-        /**
-         * @brief get the shape of ndarray.
-         *
-         * Dispatched code based on the attributes, with the following order:
-         * - T is nested array and has fixed-size
-         * - T is bounded array
-         * - fail
-         * 
-         * @return constexpr auto 
-         */
-        static constexpr auto _get()
-        {
-            // check for fixed-size array (that has tuple_size_v)
-            if constexpr (nested_array_size_v<T> > 0) {
-                // nested_array_dim_v mimics std::rank_v
-                // nested_array_size_v mimics std::extent_v
-                constexpr auto dim = nested_array_dim_v<T>;
-                auto shape = std::array<size_t,dim>{};
-                template_for<dim>([&](auto index){
-                    constexpr auto i = decltype(index)::value;
-                    using nested_t = remove_nested_array_dim_t<T,i>;
-                    std::get<i>(shape) = nested_array_size_v<nested_t>;
-                });
-                return shape;
-            }
-            // check for bounded-array (e.g. double[1][2][3]...)
-            else if constexpr (is_bounded_array_v<T>) {
-                constexpr auto rank = std::rank_v<T>;
-                auto shape = std::array<size_t,rank>{};
-                template_for<rank>([&](auto index) {
-                    constexpr auto i = decltype(index)::value;
-                    constexpr auto n = std::extent_v<T,i>;
-                    shape[i] = n;
-                });
-                return shape;
-            }
-            // fail otherwise
-            else return detail::fail_t{};
-        } // _get()
-        static inline constexpr auto value = _get();
-        using value_type = detail::fail_to_void_t<meta::remove_cvref_t<decltype(value)>>;
-    };
-
-    /**
-     * @brief helper variable template to get fixed-array size at compile-time.
-     * well-formed specialization should have `value` and `value_type`.
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline constexpr auto fixed_ndarray_shape_v = fixed_ndarray_shape<T>::value;
-
-    /**
-     * @brief check if given type T is fixed-size array
-     *
-     * Default implementation check if calls to fixed_ndarray_shape returns successful (value_type not void),
-     * returns true for such case, false otherwise.
-     * 
-     * @tparam T type to check
-     * @tparam typename customization point
-     * @see fixed_ndarray_shape
-     */
-    template <typename T, typename=void>
-    struct is_fixed_size_ndarray
-    {
-        static constexpr auto value = [](){
-            using fixed_size_t = fixed_ndarray_shape<T>;
-            using value_type   = typename fixed_size_t::value_type;
-            // TODO: use fail type instead of void
-            if constexpr (!is_void_v<value_type>)
-                return true;
-            else return false;
-        }();
-    }; // is_fixed_size_ndarray
-
-    /**
-     * @brief helper variable template to check if given type T is n-dimensional array that shape is known at compile-time
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline constexpr bool is_fixed_size_ndarray_v = is_fixed_size_ndarray<T>::value;
-
-    /**
-     * @brief Get the number of maximum size of hybrid index array
-     * 
-     * Note that this meta fn is intended to return maximum size of index/shape array,
-     * hence effectively return the maximum "dimennsion" of array.
-     * 
-     * @tparam T 
-     * @tparam typename 
-     */
-    template <typename T, typename=void>
-    struct hybrid_index_array_max_size
-    {
-        static constexpr auto value = detail::fail_t{};
-        using type = detail::fail_t;
-    }; // hybrid_index_array_max_size
-
-    template <typename T>
-    inline constexpr auto hybrid_index_array_max_size_v = hybrid_index_array_max_size<T>::value;
-
-    template <typename T, auto N, typename=void>
-    struct resize_hybrid_index_array_max_size
-    {
-        // TODO: use specific type to represent error
-        using type = void;
-    }; // resize_hybrid_ndarray_max_size
-
-    template <typename T, auto N>
-    using resize_hybrid_index_array_max_size_t = type_t<resize_hybrid_index_array_max_size<T,N>>;
-
-    template <typename T>
-    struct hybrid_ndarray_max_size
-    {
-        static constexpr auto value = detail::fail_t{};
-        using type = detail::fail_t;
-    }; // hybrid_ndarray_max_size
-
-    template <typename T>
-    inline constexpr auto hybrid_ndarray_max_size_v = hybrid_ndarray_max_size<T>::value;
-
-    template <typename T, typename=void>
-    struct is_hybrid_ndarray
-    {
-        static constexpr auto value = [](){
-            using hybrid_max_type = type_t<hybrid_ndarray_max_size<T>>;
-            if constexpr (is_same_v<detail::fail_t,hybrid_max_type>) {
-                return false;
-            } else {
-                return true;
-            }
-        }();
-    }; // is_hybrid_ndarray
-
-    template <typename T>
-    inline constexpr bool is_hybrid_ndarray_v = is_hybrid_ndarray<T>::value;
-
-    /**
-     * @brief Given hybrid ndarray and desired new maximum size, create
-     * new hybrid ndarray that has the desired new max size.
-     * 
-     * @tparam T 
-     * @tparam N desired new maximum size
-     * @tparam typename 
-     * @todo use specific error type as default, instead of void
-     */
-    template <typename T, auto N, typename=void>
-    struct resize_hybrid_ndarray_max_size
-    {
-        // TODO: use specific type to represent error
-        using type = void;
-    }; // resize_hybrid_ndarray_max_size
-
-    template <typename T, auto N>
-    using resize_hybrid_ndarray_max_size_t = type_t<resize_hybrid_ndarray_max_size<T,N>>;
-
-    namespace error
-    {
-        /**
-         * @brief Default error type for resize_hybrid_ndarray_dim
-         * 
-         */
-        struct RESIZE_HYBRID_ERROR : detail::fail_t {};
-    }
-
-    /**
-     * @brief Given hybrid ndarray and desired new dim, create
-     * new hybrid ndarray that has the desired new dim.
-     * 
-     * @tparam T 
-     * @tparam DIM desired new dimension
-     * @tparam typename 
-     */
-    template <typename T, auto DIM, typename=void>
-    struct resize_hybrid_ndarray_dim
-    {
-        using type = error::RESIZE_HYBRID_ERROR;
-    };
-
-    template <typename T, auto N>
-    using resize_hybrid_ndarray_dim_t = type_t<resize_hybrid_ndarray_dim<T,N>>;
-
-    /**
-     * @brief Check if type T is dynamic ndarray.
-     * 
-     * Default implementation check if T is fixed-size of hybrid ndarray,
-     * if both return false, then assume T is dynamic.
-     *
-     * @tparam T type to check
-     * @tparam typename 
-     */
-    template <typename T, typename=void>
-    struct is_dynamic_ndarray
-    {
-        static constexpr auto value = [](){
-            // get_element_type defined in transform, while transform
-            // depends on this file hence may create circular dependency
-            // TODO: also check get_element_type return numeric concept
-            // using element_t = get_element_type_t<T>;
-            if (is_fixed_size_ndarray_v<T> || is_hybrid_ndarray_v<T>) {
-                return false;
-            } else {
-                return true;
-            }
-        }();
-    }; // is_dynamic_ndarray
-
-    template <typename T>
-    inline constexpr bool is_dynamic_ndarray_v = is_dynamic_ndarray<T>::value;
-
-    // TODO: remove, put the logic on fixed_dim instead
-    /**
-     * @brief get fixed-array dimension
-     *
-     * Default implementation will check if type T is fixed-size ndarray
-     * and get the number of dimension from the shape then set value and value_type
-     * accordingly, otherwise value_type is void and value is fail_t{}.
-     * 
-     * @tparam T type to check
-     * @tparam typename customization poit
-     * @see is_fixed_size_ndarray
-     * @see fixed_ndarray_shape
-     * @note not to be confused with fixed_dim, in which *may* return the number of dimension for runtime array
-     */
-    template <typename T, typename=void>
-    struct fixed_ndarray_dim
-    {
-        static constexpr auto _get()
-        {
-            if constexpr (is_fixed_size_ndarray_v<T>) {
-                auto shape = fixed_ndarray_shape_v<T>;
-                auto dim   = std::tuple_size_v<decltype(shape)>;
-                return dim;
-            }
-            else return detail::fail_t{};
-        } // _get()
-        static constexpr auto value = _get();
-        using value_type = detail::fail_to_void_t<meta::remove_cvref_t<decltype(value)>>;
-    }; // fixed_ndarray_dim
-
-    /**
-     * @brief helper variable template for fixed_ndarray_dim
-     * 
-     * @tparam T 
-     */
-    template <typename T>
-    inline static constexpr auto fixed_ndarray_dim_v = fixed_ndarray_dim<T>::value;
-
-    // TODO: cleanup
-    /**
-     * @brief check if type T has fixed-dimension
-     * 
-     * @tparam T type to check
-     * @tparam typename 
-     * @see fixed_ndarray_dim
-     * @see nested_array_dim_v
-     * @note not to be confused with fixed_ndarray_dim, in which *only* return the number of dimension for fixed-size array
-     */
-    template <typename T, typename=void>
-    struct fixed_dim
-    {
-        static constexpr auto value = [](){
-            // for fixed-size ndarray, read dimension from fixed_ndarray_dim,
-            // which simply count the number of shape
-            using dim_t = fixed_ndarray_dim<T>;
-            using value_type = typename dim_t::value_type;
-            if constexpr (!is_void_v<value_type>)
-                return dim_t::value;
-            // for nested array, while the shape may only known at runtime,
-            // the dimension can be known at compile time
-            else if constexpr (nested_array_dim_v<T> > 0)
-                return nested_array_dim_v<T>;
-            else return detail::Fail;
-        }();
-        using value_type = detail::fail_to_void_t<meta::remove_cvref_t<decltype(value)>>;
-    }; // fixed_dim
-
-    /**
-     * @brief helper variable template for fixed_dim
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline static constexpr auto fixed_dim_v = fixed_dim<T>::value;
-
-    /**
-     * @brief check if type T is fixed-dimension ndarray
-     *
-     * Default implementation checks if fixed_ndarray_dim<T>
-     * has value_type that is not void.
-     * 
-     * @tparam T type to check
-     * @tparam typename customization point
-     * @see fixed_dim
-     */
-    template <typename T, typename=void>
-    struct is_fixed_dim_ndarray
-    {
-        static constexpr auto _check()
-        {
-            // for fixed-size ndarray, read dimension from fixed_dim,
-            // which simply count the number of shape
-            using dim_t = fixed_dim<T>;
-            using value_type = typename dim_t::value_type;
-            // TODO: use fail type instead of void
-            if constexpr (!is_void_v<value_type>)
-                return true;
-            else return false;
-        } // _check()
-        static constexpr auto value = _check();
-    }; // is_fixed_dim_ndarray
-
-    /**
-     * @brief helper variable template for is_fixed_dim_ndarray
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline constexpr auto is_fixed_dim_ndarray_v = is_fixed_dim_ndarray<T>::value;
-
-    /**
-     * @brief check if given type T is n-dimensional array
-     * 
-     * Default implemetnation check for distjunction of fixed-size ndarray, array1d and array2d.
-     *
-     * @tparam T type to check
-     * @tparam typename customization point
-     * @see is_fixed_size_ndarray
-     * @see is_array1d
-     * @see is_array2d
-     */
-    template <typename T, typename=void>
-    struct is_ndarray
-    {
-        static constexpr auto value = [](){
-
-            if constexpr (is_fixed_size_ndarray_v<T>)
-                return true;
-            // TODO: remove
-            else if constexpr (is_array2d_v<T>)
-                return true;
-            // TODO: remove
-            else if constexpr (is_array1d_v<T>)
-                return true;
-            else return false;
-        }();
-    }; // is_ndarray
-
-    /**
-     * @brief helper variable template to check if T is n-dimensional array
-     * 
-     * @tparam T type to check
-     */
-    template <typename T>
-    inline constexpr bool is_ndarray_v = is_ndarray<T>::value;
 
 
-    /** @} */ // end group traits
 
-} // namespace nmtools::meta
 
 #endif // NMTOOLS_META_ARRAY_HPP
